refactor: refactor model location
