+++ src/components/auth/SignUpForm.tsx
@@ -12,7 +12,7 @@ import useAccessToken from 'hooks/useAccessToken';
 import { FormProvider, useForm } from 'react-hook-form';
 import { useTranslation } from 'react-i18next';
 import { Link, useNavigate } from 'react-router-dom';
-import type { SignUpForm as SignUpFormValue } from 'models/auth//SignUpForm';
+import type { SignUpForm as SignUpFormValue } from 'models/auth/SignUpForm';
 
 type SignUpFormProps = {
   email: string;
